scripts/data-generator/requirements.txt
pyyaml==6.0
numpy
kafka-python==2.0.2